backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.99.5
schedule==1.2.0
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.0
email-validator==2.1.0
